Organise code for Organisational Rules Service tests
